[+] hw05: implement counting of set bits

+++ 05-bitboard/src/bitboard/include/bitboard/utils.h
@@ -3,6 +3,6 @@
 
 namespace bitboard {
 
-uint64_t get_moves(uint64_t bitmask);
+int get_moves(uint64_t bitmask);
 
 } // namespace bitboard
